Move global_context class from namespace dal:: to namespace dal::detail::
